Initial setup
Can now be run on Windows or via WSL.

The `ai_text` and `ai_image` scripts are more efficient when being run through the `test_plugin` script because the result is cached. As such, we don't post to OpenAI for every resolution.

Fixed some bugs

The `image_utils` now uses `pyppeteer` to render the images. This seems to be much more reliable than simply using Chrome/Chromium.